The First release of CORE MOBILE
The multifunctional mobile pocket tool
CORE Mobile is the versatile mobile application offered by MUXE B.V. which has been made available for download on iOS and Android devices. Considered the companion application to CORE Desktop, this is a “one-stop-shop” for easy and convenient access to developed products and services that have been optimised for use on mobile devices.

Cross-platform compatibility of MUXE product suite
Whilst the CORE Desktop version of CORE Browser is considered the comprehensive solution to the need for a powerful and innovative desktop workspace, CORE Mobile has been specifically designed to make available as many of these products and services within the palm of your hand.

Although a portion of the tools and services offered are not currently available for use within CORE Mobile, we strive towards integrating as many of these features as possible to provide our users with a fluid and seamless experience when utilising CORE Desktop and CORE Mobile.